Antes se me creo en el test de los metodos ya lo he añadido a la clase ControllerTest
